requirements.txt
PyYAML==5.1.2
numpy==1.16.4
tensorboard==1.15.0
tensorboard-logger==0.1.0
torch==1.6.0